# model/system/response/sys_api.go
package response

import "github.com/lxhcaicai/gin-vue-admin/server/model/system"

type SysAPIResponse struct {
	Api system.SysApi `json:"api"`
}

type SysAPIListResponse struct {
	Apis []system.SysApi `json:"apis"`
}